requirements.txt
python-telegram-bot==20.0
requests==2.26.0
python-dotenv==0.19.1
flask==2.3.0
gunicorn==22.0.0